+++ src/components/User/Assets/SystemAssets/components/AssetManager.vue
@@ -252,15 +252,6 @@
       </div>
       <!-- eslint-enable max-len -->
 
-      <div class="app__form-row">
-        <tick-field
-          class="app__form-field"
-          v-model="asset.creatorDetails.isFiat"
-          label="Fiat asset"
-          :disabled="formMixin.isDisabled"
-        />
-      </div>
-
       <div class="asset-manager-advanced__block">
         <div class="asset-manager-advanced__heading">
           <h3>Advanced</h3>
@@ -374,7 +365,6 @@ export default {
         trailingDigitsCount: '6',
         assetType: '0',
         creatorDetails: {
-          isFiat: false,
           name: '',
           logo: {},
           terms: {},
@@ -516,7 +506,6 @@ export default {
             allTasks: 0,
             creatorDetails: {
               name: this.asset.creatorDetails.name,
-              isFiat: this.asset.creatorDetails.isFiat,
               external_system_type: this.asset
                 .creatorDetails.externalSystemType,
               is_coinpayments: this.asset.creatorDetails.isCoinpayments,
@@ -537,7 +526,6 @@ export default {
             allTasks: 0,
             creatorDetails: {
               name: this.asset.creatorDetails.name,
-              isFiat: this.asset.creatorDetails.isFiat,
               external_system_type: this.asset
                 .creatorDetails.externalSystemType,
               is_coinpayments: this.asset.creatorDetails.isCoinpayments,
@@ -557,7 +545,7 @@ export default {
         Bus.$emit('recheckConfig')
 
         this.$store.dispatch('SET_INFO', 'Submitted successfully.')
-        this.$router.push({ name: 'assets.masterAssets.index' })
+        this.$router.push({ name: 'assets.systemAssets.index' })
       } catch (error) {
         ErrorHandler.process(error)
       }